Renamed module name and fixed setup file
