+++ setup.cfg
@@ -33,7 +33,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    readit = pygenpass.__init__:main
+    pygenpass = pygenpass.__init__:main
 
 [flake8]
 max_line_length = 100
